+++ examples/deferred.rs
@@ -36,7 +36,7 @@ fn main() {
     let gl = gl::Gl::load_with(|s| video_ctx.gl_get_proc_address(s) as *const std::os::raw::c_void);
 
     // Renderer
-    let renderer = renderer::Pipeline::create(&gl, width, height).unwrap();
+    let renderer = pipeline::DeferredPipeline::create(&gl, width, height).unwrap();
 
     // Scene
     let mut scene = scene::Scene::create();
@@ -75,7 +75,7 @@ fn main() {
         }
 
         // draw
-        renderer.deferred_pass(&camera, &scene).unwrap();
+        renderer.draw(&camera, &scene).unwrap();
 
         window.gl_swap_window();
     };
