+++ src/renderer.rs
@@ -1,86 +1,3 @@
-use camera;
-use scene;
-use input;
-use glm;
-use gl;
-use num_traits::identities::One;
-use screen;
-use core::rendertarget;
-use core::rendertarget::Rendertarget;
-
-#[derive(Debug)]
-pub enum Error {
-    Scene(scene::Error),
-    Screen(screen::Error),
-    Rendertarget(rendertarget::Error)
-}
-
-impl From<scene::Error> for Error {
-    fn from(other: scene::Error) -> Self {
-        Error::Scene(other)
-    }
-}
-
-impl From<screen::Error> for Error {
-    fn from(other: screen::Error) -> Self {
-        Error::Screen(other)
-    }
-}
-
-impl From<rendertarget::Error> for Error {
-    fn from(other: rendertarget::Error) -> Self {
-        Error::Rendertarget(other)
-    }
-}
-
-pub struct Pipeline {
-    gl: gl::Gl,
-    screen: screen::Screen,
-    geometry_pass_rendertarget: rendertarget::ColorRendertarget
-}
-
-
-impl Pipeline
-{
-    pub fn create(gl: &gl::Gl, width: usize, height: usize) -> Result<Pipeline, Error>
-    {
-        let screen = screen::Screen::create(gl, width, height)?;
-        let geometry_pass_rendertarget = rendertarget::ColorRendertarget::create(&gl, width, height)?;
-        Ok(Pipeline { gl: gl.clone(), screen, geometry_pass_rendertarget })
-    }
-
-    pub fn forward_pass(&self, camera: &camera::Camera, scene: &scene::Scene) -> Result<(), Error>
-    {
-        let input = input::ReflectingInput { model: glm::Matrix4::one(),view: camera.get_view(), projection: camera.get_projection(),
-            camera_position: camera.position };
-
-        self.screen.rendertarget.bind();
-        self.screen.rendertarget.clear();
-
-        scene.draw(&input)?;
-        Ok(())
-    }
-
-    pub fn deferred_pass(&self, camera: &camera::Camera, scene: &scene::Scene) -> Result<(), Error>
-    {
-        let reflecting_input = input::ReflectingInput { model: glm::Matrix4::one(),view: camera.get_view(), projection: camera.get_projection(),
-            camera_position: camera.position };
-
-        self.geometry_pass_rendertarget.bind();
-        self.geometry_pass_rendertarget.clear();
-
-        scene.draw(&reflecting_input)?;
-
-        self.screen.rendertarget.bind();
-        self.screen.rendertarget.clear();
-
-        let emitting_input = input::EmittingInput{ camera_position: camera.position,
-            color_texture: self.geometry_pass_rendertarget.color_texture.clone()};
-
-        scene.shine_lights(&emitting_input)?;
-        Ok(())
-    }
-}
 
 #[cfg(target_os = "emscripten")]
 pub fn set_main_loop<F>(main_loop: F) where F: FnMut()
